requirements.txt
streamlit
python-dotenv
openai
langchain
matplotlib
plotly